# api-trello.swagger.yaml
openapi: 3.0.4
info:
  title: Trello API Documentation
  description: |-
    This backend project is designed to support a Trello-like task management system, handling user authentication, board and list management, task assignments, and real-time updates.
    Some useful links:
    - [Trello API repository](https://github.com/minhuy206/trello-api)
  contact:
    email: forwork@minhuy.dev
  version: 1.0.0
servers:
  - url: http://localhost:8017/v1
    description: Local server
  - url: https://api-trello.minhuy.dev/v1
    description: Production server
...
paths:
...
  /users/send-otp:
    post:
      tags:
        - User
      summary: Send OTP
      description: Send OTP to the user. This could be used for registration.
      operationId: sendOtp
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOTPBody"
        required: true
      responses:
        '201':
          description: OTP sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: OTP has been sent to your email    
        '422':
          description: Validation exception
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnprocessableEntityError"
  /users/refresh-token:
    post:
      tags:
        - User
      summary: Refresh access token
      description: Refresh access token using refresh token.
      operationId: refreshToken
      requestBody:
        content:
          application/json:
            schema:
              type: object
        required: true
      responses:
        '201':
          description: Refresh token successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        '422':
          description: Validation exception
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnprocessableEntityError"
      security:
        - CookieAuth: []
  /users/forgot-password:
    post:
      tags:
        - User
      summary: Forgot password
      description: Send an email to password.
      operationId: forgotPassword
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordBody"
        required: true
      responses:
        '201':
          description: An email has been sent to reset password
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Reset password link has been sent to your email
        '422':
          description: Validation exception
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnprocessableEntityError"
  /users/reset-password:
    post:
      tags:
        - User
      summary: Reset password
      description: Reset password using OTP.
      operationId: resetPassword
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordBody"
        required: true
      responses:
        '201':
          description: Password has been changed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Password has been changed successfully
        '422':
          description: Validation exception
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnprocessableEntityError"
  /users/verify:
    put:
      tags:
        - User
      summary: Verify user
      description: Verify user with OTP.
      operationId: verifyUser
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyBody"
        required: true
      responses:
        '201':
          description: User verified successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /invitations:
    post:
      tags:
        - Invitation
      summary: Create a new invitation
      description: Create a new invitation.
      operationId: createInvitation
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvitationBody"
        required: true
      responses:
        '201':
          description: Create invitation successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invitation"
        '422':
          description: Validation exception
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnprocessableEntityError"
      security:
        - BearerAuth: []
    get:
      tags:
        - Invitation
      summary: Get all invitations
      description: Get all invitations.
      operationId: getAllInvitations
      parameters: []
      responses:
        '200':
          description: Get all invitations successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  invitations:
                    type: array
                    items:
                      $ref: "#/components/schemas/Invitation"
                  totalInvitations:
                    type: integer
                    example: 10
      security:
        - BearerAuth: []
...
components:
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
          format: ObjectId
          example: 67f36ac75d3320bd988348b5
        email:
          type: string
          example: email@domain.com
        username:
          type: string
          example: username
        displayName:
          type: string
          example: 'Your display name'
        avatar:
          type: string
          example: https://example.com/avatar.jpg
        role:
          $ref: "#/components/schemas/Role"
        isVerified:
          type: boolean
          example: true   
        createdAt:
          type: string
          format: ISO8601
          example: '2023-10-01T12:00:00Z'
        updatedAt:
          type: string
          format: ISO8601
          example: '2023-10-01T12:00:00Z'
        _destroy:
          type: boolean
          example: false
...
    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
          example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c'
        refreshToken:
          type: string
          example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c'
        user: 
          $ref: "#/components/schemas/User"
...
    VerifyBody:
      type: object
      properties:
        email:
          type: string
          example: email@domain.com
        otp:
          type: string
          example: '123456'
    SendOTPBody:
      type: object
      properties:
        email: 
          type: string
          example: email@domain.com
    ForgotPasswordBody: 
      type: object
      required:
        - email | username
      properties:
        email:
          type: string
          example: email@domain.com
        username: 
          type: string
          example: username
    ResetPasswordBody:
      type: object
      properties:
        email:
          type: string
          example: '12345678a'
        password:
          type: string
          example: '12345678a'
        confirmPassword:
          type: string
          example: '12345678a'
        token:
          type: string
          format: uuid
          example: '123e4567-e89b-12d3-a456-426614174000'
...
    UnprocessableEntityError:
        type: object
        properties:
          statusCode:
            type: integer
            format: int32
            example: 422
          message:
            type: string
            example: Validation error
          error:
            type: string
            example: Unprocessable Entity
    Invitation:
        type: object
        properties:
          boardId:
            type: string
            format: ObjectId
            example: 67f36ac75d3320bd988348b5
          invitee:
            $ref: "#/components/schemas/User"
          status:
            $ref: "#/components/schemas/InvitationStatus"
          createdBy:
            $ref: "#/components/schemas/User"
          updateAt:
            type: string
            format: ISO8601
            example: '2023-10-01T12:00:00Z'
          createdAt:
            type: string
            format: ISO8601
            example: '2023-10-01T12:00:00Z'
    CreateInvitationBody:
        type: object
        properties:
          boardId:
            type: string
            format: ObjectId
            example: 67f36ac75d3320bd988348b5
          inviteeEmail:
            type: string
            format: email
            example: email@domain.com
...
    Role:
      type: string
      enum: [admin, client]
    InvitationStatus:
      type: string
      enum: [accepted, pending, rejected]
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth: 
      type: apiKey
      name: Cookie
      in: header